requirements.txt
openai>=1.75.0
pandas>=2.2.0
python-dotenv>=1.0.1
pydantic>=2.7.0
tqdm>=4.66.0
pypdf>=4.2.0
python-docx>=1.1.2
openpyxl>=3.1.2